% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.glmb.R
\name{anova.glmb}
\alias{anova.glmb}
\title{Analysis of Deviance for Bayesian Generalized Linear Model Fits}
\usage{
\method{anova}{glmb}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glmb}, typically the result of a call to \link{glmb}}

\item{\ldots}{Other arguments passed to or from other methods.}
}
\value{
An object of class \code{"anova"} inheriting from class \code{"data.frame"}.
}
\description{
Compute an analysis of deviance table for one (current implementation) or more (future)
Bayesian generalized linear model fits. The structure follows the sequential analysis
of deviance \insertCite{McCullagh1989}{nmathopencl}, with Bayesian extensions for DIC, pD,
Mahalanobis shift, and directional tail probability \insertCite{Spiegelhalter2002}{nmathopencl}.
}
\details{
Specifying a single object (currently only implementation) gives a sequential
analysis of deviance table for that fit. The reductions in residual deviance as each
term of the formula is added in turn are given as the rows of a table, plus the residual
deviances themselves. The Mahalanobis shift and pDirectional columns report prior-posterior
disagreement diagnostics; see \code{\link{directional_tail}} and
\insertCite{glmbayesChapterA04}{nmathopencl}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{directional_tail}}, \code{\link{summary.glmb}}, \code{\link{glmb}},
\code{\link{nmathopencl-package}}; \code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{anova.glm}}
}
