% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitter_functions.R
\name{rNormal_reg.wfit}
\alias{rNormal_reg.wfit}
\alias{glmb.wfit}
\title{Bayesian Weighted Fitting Engines}
\usage{
rNormal_reg.wfit(
  x,
  y,
  P,
  mu,
  w,
  offset = NULL,
  method = "qr",
  tol = 1e-07,
  singular.ok = TRUE,
  ...
)

glmb.wfit(
  x,
  y,
  weights = rep.int(1, nobs),
  offset = rep.int(0, nobs),
  family = gaussian(),
  Bbar,
  P,
  betastar,
  method = "qr",
  tol = 1e-07,
  singular.ok = TRUE,
  ...
)
}
\arguments{
\item{x}{design matrix of dimension \code{n * p}.}

\item{y}{vector of observations of length \code{n}, or a matrix with
    \code{n} rows.}

\item{P}{Prior precision matrix of dimension \code{p * p}.}

\item{mu}{Prior mean vector of length \code{p}.}

\item{w}{vector of weights (length \code{n}) to be used in the fitting
    process for the \code{wfit} functions.  Weighted least squares is
    used with weights \code{w}, i.e., \code{sum(w * e^2)} is minimized.}

\item{offset}{(numeric of length \code{n}).  This can be used to
    specify an \emph{a priori} known component to be included in the
    linear predictor during fitting.}

\item{method}{currently, only \code{method = "qr"} is supported.}

\item{tol}{tolerance for the \code{\link{qr}} decomposition.  Default
    is 1e-7.}

\item{singular.ok}{logical. If \code{FALSE}, a singular model is an
    error.}

\item{...}{currently disregarded.}

\item{weights}{an optional vector of \emph{prior weights} to be used in the fitting process.
Should be \code{NULL} or a numeric vector.}

\item{family}{a description of the error distribution and link function to be used in the model.
Should be a family function. (see \code{\link{family}} for details of family functions.)}

\item{Bbar}{Prior mean vector of length \code{p}.}

\item{betastar}{Posterior mode vector of length \code{p} which has already been estimated.}
}
\value{
a \code{\link{list}} with components:

a \code{\link{list}} wih components:
}
\description{
These functions provide the Bayesian analogue of \code{lm.wfit}. They implement the core weighted least squares
step used inside Bayesian linear linear models, incorporating prior
precision and posterior mode information.
}
\details{
\code{rNormal_reg.wfit} performs the Bayesian weighted least squares update
for linear models under a Normal prior.

\code{glmb.wfit} performs the corresponding update for generalized linear
models, reconstructing the weighted least squares step using the posterior
mode and the GLM family functions.
}
