% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case.names.glmb.R
\name{case.names.glmb}
\alias{case.names.glmb}
\alias{variable.names.glmb}
\title{Case and Variable Names of Fitted Models}
\usage{
\method{case.names}{glmb}(object, full = FALSE, ...)

\method{variable.names}{glmb}(object, full = FALSE, ...)
}
\arguments{
\item{object}{a fitted model object, typically of class \code{"glmb"}.}

\item{full}{logical; if \code{TRUE}, all names (including zero weights, ...) are returned.}

\item{\ldots}{further arguments passed to or from other methods.}
}
\value{
A character vector
}
\description{
Simple utilities returning (non-missing) case names, and (non-eliminated) variable names.
}
\seealso{
\code{\link{glmb}}, \code{\link{nmathopencl-package}}; \code{\link{rglmb}}, \code{\link{rlmb}},
\code{\link{lmb}}; \code{\link[stats]{case.names}}.
}
