% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.glmb.R
\name{confint.glmb}
\alias{confint.glmb}
\title{Credible Intervals for Model Parameters}
\usage{
\method{confint}{glmb}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted model object of class \code{"glmb"}. Typically the result of a call to \link{glmb}.}

\item{parm}{a specification (not yet implemented) of which parameters are to be given credible sets,
either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{the credible interval required.}

\item{\ldots}{additional argument(s) for methods.}
}
\value{
A matrix (or vector) with columns giving lower and
upper credible limits for each parameter. These will be labeled
(1-level)/2 and 1-(1-level)/2 in \\% (by default 2.5\\% and 97.5\\%).
}
\description{
Computes Bayesian credible intervals for model parameters from posterior draws.
Intervals are constructed as quantiles of the posterior distribution, following
the standard approach for posterior summaries \insertCite{Gelman2013}{nmathopencl}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{summary.glmb}}, \code{\link{vcov.glmb}}, \code{\link{glmb}},
\code{\link{nmathopencl-package}}; \code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{confint}} for classical confidence intervals
}
