% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.rglmb.R
\name{deviance.rglmb}
\alias{deviance.rglmb}
\title{Model Deviance}
\usage{
\method{deviance}{rglmb}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"rglmb"}, typically the result of a call to \link{rglmb}}

\item{...}{further arguments to or from other methods}
}
\value{
A vector with the deviance extracted from the \code{object}.
}
\description{
Returns the deviance of a fitted Bayesian Generalized Linear Model
}
\seealso{
\code{\link{rglmb}}; \code{\link{glmb}}, \code{\link{nmathopencl-package}};
\code{\link{rlmb}}, \code{\link{lmb}}; \code{\link[stats]{deviance}}.
}
