% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.coef.glmb.R
\name{dummy.coef.glmb}
\alias{dummy.coef.glmb}
\alias{print.dummy_coef.glmb}
\title{Extract Coefficients in Original Coding}
\usage{
\method{dummy.coef}{glmb}(object, use.na = FALSE, ...)

\method{print}{dummy_coef.glmb}(x, ...)
}
\arguments{
\item{object}{a \code{glmb} model fit}

\item{use.na}{logical flag for coefficients in a singular model. If use.na is true,
undetermined coefficients will be missing; if false they will get one possible value.}

\item{x}{object to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\value{
A list giving for each term the draws for the coefficients.
}
\description{
Extracts coefficients in terms of the original factor levels rather than the
coded variables. The logic mirrors \code{\link[stats]{dummy.coef}} for \code{lm}
objects \insertCite{Chambers1992,VenablesRipley2002}{nmathopencl}.
}
\details{
A fitted linear model has coefficients for the contrasts of the factor terms,
usually one less in number than the number of levels. This function re-expresses the
coefficients in the original coding; as the coefficients will have been fitted in the
reduced basis, any implied constraints (e.g., zero sum for contr.helmert or contr.sum)
will be respected. There will be little point in using dummy.coef for contr.treatment
contrasts, as the missing coefficients are by definition zero.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{summary.glmb}}, \code{\link{glmb}}, \code{\link{nmathopencl-package}};
\code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{dummy.coef}}
}
