% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_glmbayes.R
\name{Ex_glmb_Standardize_Model}
\alias{Ex_glmb_Standardize_Model}
\title{Standardize A Non-Gaussian Model}
\usage{
Ex_glmb_Standardize_Model(y, x, P, bstar, A1)
}
\arguments{
\item{y}{a vector of observations of length m}

\item{x}{a design matrix of dimension m*p}

\item{P}{Positive-definite prior precision (\eqn{m \\times m}).}

\item{bstar}{a matrix containing the posterior mode from an optimization step}

\item{A1}{a matrix containing the posterior precision matrix at the posterior mode}
}
\value{
A list with the following components
\item{bstar2}{Standardized Posterior Mode}
\item{A}{Standardized Data Precision Matrix}
\item{x2}{Standardized Design Matrix}
\item{mu2}{Standardized Prior Mean vector}
\item{P2}{Standardized Precision Matrix Added to log-likelihood}
\item{L2Inv}{A matrix used when undoing the first step in standardization described below}
\item{L3Inv}{A matrix used when undoing the second step in standardization described below}
}
\description{
Standardizes a Non-Gaussian Model prior to Envelope Creation
}
\details{
Starts from the posterior mode quantities and proceeds in three transformations.\cr
Step 1: posterior-precision eigendecomp.\cr Interim model gets identity posterior precision.\cr
Step 2: isolate diagonal \code{epsilon}; remainder behaves like likelihood information.\cr
Step 3: another eigendecomp diagonalizes data precision \code{A}.\cr The
prior tied to \code{epsilon} becomes identity.\cr
\insertCite{Nygren2006}{nmathopencl}
}
\references{
\insertAllCited{}
}
