% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDIC.R
\name{extractAIC.glmb}
\alias{extractAIC.glmb}
\alias{extractAIC.rglmb}
\alias{extractDIC}
\title{Extract DIC from a Fitted Bayesian Model}
\usage{
\method{extractAIC}{glmb}(fit, ...)

\method{extractAIC}{rglmb}(fit, ...)

extractDIC(fit, ...)
}
\arguments{
\item{fit}{A fitted model of class \code{"glmb"} or \code{"rglmb"}.}

\item{...}{Additional arguments passed to or from methods.}
}
\value{
A named numeric vector with components:
\describe{
\item{pD}{Estimated effective number of parameters}
\item{DIC}{Deviance Information Criterion}
}
}
\description{
Computes the Deviance Information Criterion (DIC) and the effective
number of parameters for Bayesian generalized linear models fitted
via \code{glmb()} or \code{rglmb()}. The DIC, introduced by
\insertCite{Spiegelhalter2002}{nmathopencl}, provides a Bayesian analog to the
AIC \insertCite{Akaike1974}{nmathopencl}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{summary.glmb}}, \code{\link{glmb}}, \code{\link{nmathopencl-package}},
\code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{extractAIC}} for the classical AIC computation on \code{lm}/\code{glm} fits
}
