% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.glmb.R
\name{glmb.influence.measures}
\alias{glmb.influence.measures}
\alias{rstandard.glmb}
\alias{rstudent.glmb}
\alias{glmb.dffits}
\alias{dfbetas.glmb}
\alias{glmb.covratio}
\alias{cooks.distance.glmb}
\title{Bayesian Regression Diagnostics}
\usage{
glmb.influence.measures(model, infl = influence(model))

\method{rstandard}{glmb}(model, ..., infl = influence(model))

\method{rstudent}{glmb}(model, ..., infl = influence(model))

glmb.dffits(model, infl = influence(model))

\method{dfbetas}{glmb}(model, ..., infl = influence(model))

glmb.covratio(model, infl = influence(model))

\method{cooks.distance}{glmb}(model, ..., infl = influence(model))
}
\arguments{
\item{model}{an \R object, typically returned by \code{\link[stats]{lm}} or
    \code{\link[stats]{glm}}.}

\item{infl}{influence structure as returned by \code{influence.glmb}
(the latter only for the glm method of \code{rstudent} and \code{cooks.distance}).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a \code{\link{list}} with components as returned by the underlying \pkg{stats} functions.
}
\description{
These functions compute Cook's distance, dfbetas, dffits, covratio, and standardized/studentized
residuals for fitted Bayesian GLM objects. They delegate to the corresponding \pkg{stats} functions
applied to the fitted GLM component.
}
