% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmbayes-package.R
\docType{package}
\name{glmbayes-package}
\alias{glmbayes-package}
\title{glmbayes: Bayesian Generalized Linear Models (IID Samples)}
\description{
Generates iid samples for Bayesian Generalized Linear Models.
}
\author{
\strong{Maintainer}: Kjell Nygren \email{kjell.a.nygren@gmail.com}

}
\keyword{internal}
