% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.glmb.R
\name{logLik.glmb}
\alias{logLik.glmb}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{glmb}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glmb}, typically the result of a call to \link{glmb}}

\item{...}{further arguments to or from other methods}
}
\value{
The function returns a vector, \code{logLikout} with the estimated log-likelihood for each draw.
}
\description{
This function is a method function for the \code{"glmb"} class used to
Extract the log-Likelihood from a Bayesian Generalized Linear Model.
}
\seealso{
\code{\link{glmb}}, \code{\link{nmathopencl-package}}; \code{\link{rglmb}}, \code{\link{rlmb}},
\code{\link{lmb}}; \code{\link[stats]{logLik}}.
}
