% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{pbeta_opencl}
\alias{pbeta_opencl}
\title{OpenCL-backed pbeta linkage check}
\usage{
pbeta_opencl(n, x, a, b, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed pbeta linkage check
}
