% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{pnchisq_opencl}
\alias{pnchisq_opencl}
\title{OpenCL-backed pnchisq linkage check}
\usage{
pnchisq_opencl(n, x, df, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed pnchisq linkage check
}
