% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{pnf_opencl}
\alias{pnf_opencl}
\title{OpenCL-backed pnf linkage check}
\usage{
pnf_opencl(n, x, df1, df2, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed pnf linkage check
}
