% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{qbeta_opencl}
\alias{qbeta_opencl}
\title{OpenCL-backed qbeta linkage check}
\usage{
qbeta_opencl(n, p, a, b, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qbeta linkage check
}
