% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{qbinom_opencl}
\alias{qbinom_opencl}
\title{OpenCL-backed qbinom linkage check}
\usage{
qbinom_opencl(n, p, size, prob, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qbinom linkage check
}
