% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_opencl.R
\name{qchisq_opencl}
\alias{qchisq_opencl}
\title{OpenCL-backed qchisq linkage check}
\usage{
qchisq_opencl(n, p, df, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qchisq linkage check
}
