% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_linkage_opencl.R
\name{qnbinom_mu_opencl}
\alias{qnbinom_mu_opencl}
\title{OpenCL-backed qnbinom_mu linkage check}
\usage{
qnbinom_mu_opencl(n, p, size, mu, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qnbinom_mu linkage check
}
