% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{qnchisq_opencl}
\alias{qnchisq_opencl}
\title{OpenCL-backed qnchisq linkage check}
\usage{
qnchisq_opencl(n, p, df, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qnchisq linkage check
}
