% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{qnf_opencl}
\alias{qnf_opencl}
\title{OpenCL-backed qnf linkage check}
\usage{
qnf_opencl(n, p, df1, df2, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qnf linkage check
}
