% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{qpois_opencl}
\alias{qpois_opencl}
\title{OpenCL-backed qpois linkage check}
\usage{
qpois_opencl(n, p, lambda, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed qpois linkage check
}
