% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_runtime_opencl.R
\name{r_pow_di_opencl}
\alias{r_pow_di_opencl}
\title{OpenCL-backed R_pow_di linkage check}
\usage{
r_pow_di_opencl(n, x, n_exp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed R_pow_di linkage check
}
