% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{r_pow_opencl}
\alias{r_pow_opencl}
\title{OpenCL-backed R_pow linkage check}
\usage{
r_pow_opencl(n, x, y, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed R_pow linkage check
}
