% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{r_unif_index_opencl}
\alias{r_unif_index_opencl}
\title{OpenCL-backed R_unif_index linkage check}
\usage{
r_unif_index_opencl(n, dn, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed R_unif_index linkage check
}
