% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.glmb.R
\name{residuals.glmb}
\alias{residuals.glmb}
\alias{residuals.rglmb}
\alias{residuals.lmb}
\title{Accessing Bayesian Generalized Linear Model Fits}
\usage{
\method{residuals}{glmb}(object, ysim = NULL, ...)

\method{residuals}{rglmb}(object, ysim = NULL, ...)

\method{residuals}{lmb}(object, ysim = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{glmb}, typically the result of a call to \link{glmb}}

\item{ysim}{Optional simulated data for the data y.}

\item{\ldots}{further arguments to or from other methods}
}
\value{
A matrix \code{DevRes} of dimension \code{n} times \code{p} containing
the Deviance residuals for each draw. If ysim is provided, the residuals are based
on a comparison to the simulated data instead. The credible intervals
for residuals based on simulated data should be a more appropriate measure of
whether individual residuals represent outliers or not.
}
\description{
Extract deviance residuals from fitted Bayesian GLM objects. The residuals
use the family's deviance residuals function as in \code{\link[stats]{residuals.glm}}
\insertCite{McCullagh1989}{nmathopencl}.
}
\details{
These functions are all \link{methods} for class \code{glmb}, \code{lmb}, or \code{summary.glmb} objects.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.glmb}}, \code{\link{summary.glmb}}, \code{\link{glmb}},
\code{\link{nmathopencl-package}}; \code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{residuals.glm}}
}
