% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmath_opencl_linkage.R
\name{rpois_opencl}
\alias{rpois_opencl}
\title{OpenCL-backed rpois linkage check}
\usage{
rpois_opencl(n, lambda, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed rpois linkage check
}
