% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.glmb.R
\name{simulate.glmb}
\alias{simulate.glmb}
\title{Simulate Responses}
\usage{
\method{simulate}{glmb}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{glmb}, typically the result of a call to the
function \code{glmb}.}

\item{nsim}{Defunct (see below).}

\item{seed}{an object specifying if and how the random number generator should be
initialized (seeded).}

\item{\ldots}{Additional arguments passed to the function. Will
frequently include a matrix pred of simulated predictions from
the predict function, the family (e.g., binomial) and an optional
vector of weights specifying prior.weights for the simulated values (default is 1)}
}
\value{
Simulated values for data corresponding to simulated model predictions that correspond either
to the original data or to a \code{newdata} data frame provided to the predict function.
}
\description{
Simulate responses from the posterior predictive distribution corresponding to a
fitted \code{glmb} object \insertCite{Gelman2013}{nmathopencl}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{predict.glmb}}, \code{\link{glmb}}, \code{\link{nmathopencl-package}};
\code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{simulate}} (e.g. \code{simulate.glm}, \code{simulate.lm} for classical fits);
see \insertCite{glmbayesChapter04}{nmathopencl} for model statistics.
}
