% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationpipeline.R
\name{SimulationPipeline}
\alias{SimulationPipeline}
\title{Low-Level Simulation Pipeline for Bayesian GLMs}
\description{
A detailed overview of the low-level simulation pipeline used by
\code{rglmb()} and related functions. These routines implement the
optimization -> standardization -> envelope sizing -> envelope construction ->
sampling -> back-transformation workflow described in \insertCite{Nygren2006}{nmathopencl}.
}
\details{
(summaries of each step)
}
\references{
\insertAllCited{}
}
