% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rglmb.R
\name{summary.rglmb}
\alias{summary.rglmb}
\alias{print.summary.rglmb}
\title{Summarizing Bayesian Generalized Linear Model Distribution Functions}
\usage{
\method{summary}{rglmb}(object, ...)

\method{print}{summary.rglmb}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"rglmb"} for which a
summary is desired.}

\item{x}{an object of class \code{"summary.rglmb"} for which a printed output is desired.}

\item{digits}{the number of significant digits to use when printing.}

\item{\ldots}{Additional optional arguments}
}
\value{
\code{summary.rglmb} returns a object of class \code{"summary.rglmb"}, a
list with components:
\item{n}{number of draws generated}
\item{coefficients1}{Matrix with the prior mean and approximate weight for the prior relative to the data}
\item{coefficients}{Matrix with columns for the posterior mode, posterior mean, posterior standard
deviation, monte carlo error, and tail probabilities (posterior probability of observing a
value for the coefficient as extreme as the prior mean)}
\item{Percentiles}{Matrix with estimated percentiles associated with the posterior density}
}
\description{
These functions are all \code{\link{methods}} for class \code{rglmb} or \code{summary.rglmb} objects.
}
\details{
The \code{summary.rglmb} function summarizes the output from the
\code{\link{rglmb}} function. It takes an object of class \code{rglmb} as an input.
The output to a large extent mirrors the output from the \code{\link{summary.glm}}
function. This is particularly true for the print output from the function
(i.e. the output of the function \code{\link{print.summary.rglmb}}).
}
\seealso{
\code{\link{rglmb}}; \code{\link{glmb}}, \code{\link{nmathopencl-package}},
\code{\link{rlmb}}, \code{\link{lmb}}, \code{\link{summary}},
\code{\link[stats]{summary.lm}}, \code{\link[stats]{summary.glm}}.
}
