% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.glmb.R
\name{vcov.glmb}
\alias{vcov.glmb}
\title{Calculate Variance-Covariance Matrix for a Fitted Model Object}
\usage{
\method{vcov}{glmb}(object, ...)
}
\arguments{
\item{object}{fitted model object, typically the result of a call to \code{\link{glmb}}.}

\item{\ldots}{additional arguments for method functions.}
}
\value{
A matrix of estimated covariances between the parameter estimates
in the linear or non-linear predictor of the model. This should have
row and column names corresponding to the parameter names given by the
\code{\link{coef}} method.
}
\description{
Returns the posterior variance-covariance matrix of the regression coefficients
from a fitted Bayesian GLM object \insertCite{Gelman2013}{nmathopencl}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{confint.glmb}}, \code{\link{summary.glmb}}, \code{\link{glmb}},
\code{\link{nmathopencl-package}}; \code{\link{rglmb}}, \code{\link{rlmb}}, \code{\link{lmb}};
\code{\link[stats]{vcov}}
}
