% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_opencl.R
\name{pchisq_opencl}
\alias{pchisq_opencl}
\title{OpenCL-backed pchisq linkage check}
\usage{
pchisq_opencl(n, x, df, ncp, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed pchisq linkage check
}
