% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{Prior_Setup}
\alias{Prior_Setup}
\title{Setup Prior Objects}
\usage{
Prior_Setup(
  formula,
  data = NULL,
  family = gaussian(),
  pwt = 0.01,
  n_prior = NULL,
  sd = NULL,
  intercept_source = c("null_model", "full_model"),
  effects_source = c("null_effects", "full_model"),
  mu = NULL,
  subset = NULL,
  na.action = na.fail,
  drop.unused.levels = FALSE,
  xlev = NULL,
  ...
)
}
\arguments{
\item{formula}{a model \code{\link[stats]{formula}} or \code{\link[stats]{terms}}
    object or an \R object.}

\item{data}{a data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame),
    containing the variables in \code{formula}.  Neither a matrix nor an
    array will be accepted.}

\item{family}{a description of the error distribution and linke function to be used in the model.}

\item{pwt}{Weight on the prior relative to the likelihood function at the the maximum likelihood estimate.}

\item{n_prior}{Optional argument with number of prior observations (either a scalar or a vector). When provided, this used together with the number of likelihood observations to compute the pwt.}

\item{sd}{Optional vector argument with the prior standard deviations for the coefficients}

\item{intercept_source}{Specifies the method through which the prior mean for the intercept term is set. Options are based on the null intercept only model (null_model) or full_models. The default is the null model which is safer if variables are not centered.}

\item{effects_source}{Specifies the method through which the prior means for the effects terms are set. Options are null_effects (prior means set to zero) or full_model (effect means set to match maximum likelihood estimates).}

\item{mu}{Optional vector argument with the prior means for the coefficients}

\item{subset}{a specification of the rows/observations to be used:
    defaults to all. This can be any valid indexing vector (see
    \code{\link{[.data.frame}}) for the rows of \code{data}, or
    a (logical) expression using variables in \code{data} or
    if that is not supplied, in
    \code{formula}.  (See additional details about how this argument
    interacts with data-dependent bases under \sQuote{Details} below.)}

\item{na.action}{how \code{NAs} are treated. The default is first, any \code{\link{na.action}} attribute of 
data, second a \code{na.action} setting of \link{options}, and third \code{na.fail} if that is unset. 
The \code{factory-fresh} default is \code{na.omit}. Another possible value is \code{NULL}.}

\item{drop.unused.levels}{should factors have unused levels dropped?
    Defaults to \code{FALSE}.}

\item{xlev}{a named list of character vectors giving the full set of levels
    to be assumed for each factor.}

\item{...}{for \code{model.frame} methods, a mix of further
    arguments such as \code{data}, \code{na.action}, \code{subset} to pass
    to the default method.  Any additional arguments (such as
    \code{offset} and \code{weights} or other named arguments) which
    reach the default method are used to create further columns in the
    model frame, with parenthesised names such as \code{"(offset)"}.

    For \code{get_all_vars}, further named columns to include
    in the model frame.}
}
\value{
A list with items related to the prior.
\item{mu}{A prior mean vector}
\item{Sigma}{A prior variance-covariance matrix}
\item{dispersion}{Empirical bayes estimate for the dispersion (gaussian model only)}
\item{model}{The model frame from \code{object} if it exists}
\item{x}{The design matrix from \code{object} if it exists}
}
\description{
Sets up the structure for the Prior mean and Variance Matrices using information from a classical model.
}
\examples{
## Dobson (1990) Page 93: Randomized Controlled Trial :
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
print(d.AD <- data.frame(treatment, outcome, counts))
## Step 1: Set up Prior
ps=Prior_Setup(counts ~ outcome + treatment)
mu=ps$mu
V=ps$Sigma
print(ps)
}
\seealso{
Other prior: 
\code{\link{Prior_Check}()}
}
\concept{prior}
