% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rext_utils_opencl.R
\name{r_check_user_interrupt_opencl}
\alias{r_check_user_interrupt_opencl}
\title{OpenCL-backed R_CheckUserInterrupt linkage check}
\usage{
r_check_user_interrupt_opencl(n, fallback = TRUE, verbose = FALSE)
}
\description{
OpenCL-backed R_CheckUserInterrupt linkage check
}
