% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunction.R
\name{simfunction}
\alias{simfunction}
\alias{simfunction.default}
\alias{print.simfunction}
\title{Extract Simulation Function Metadata}
\usage{
simfunction(object, ...)

\method{simfunction}{default}(object, ...)

\method{print}{simfunction}(x, ...)
}
\arguments{
\item{object}{A fitted model object containing a pfamily component.}

\item{...}{Additional arguments (currently unused).}

\item{x}{an object, a Simulation function that is to be printed}
}
\value{
An object of class `"simfunction"` containing the function name, call, and arguments.
}
\description{
Returns a structured summary of the simulation function used in a model object.
}
