% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rgamma_reg.R
\name{summary.rGamma_reg}
\alias{summary.rGamma_reg}
\alias{print.summary.rGamma_reg}
\title{Summarizing Bayesian gamma_reg Distribution Functions}
\usage{
\method{summary}{rGamma_reg}(object, ...)

\method{print}{summary.rGamma_reg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{an object of class \code{"rGamma_reg"} for which a
summary is desired.}

\item{x}{an object of class \code{"summary.rGamma_reg"} for which a printed output is desired.}

\item{digits}{the number of significant digits to use when printing.}

\item{\ldots}{Additional optional arguments}
}
\description{
These functions are all \code{\link{methods}} for class \code{rgamna_reg} or \code{summary.rgamma_reg} objects.
}
